# cmd/api/admin-api/docs/swagger.yaml
basePath: /
definitions:
  apiresp.ApiResponse:
    properties:
      data: {}
      errCode:
        type: integer
      errDlt:
        type: string
      errMsg:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.AddDefaultFriendReq:
    properties:
      userIDs:
        items:
          type: string
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.AddDefaultGroupReq:
    properties:
      groupIDs:
        items:
          type: string
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.AddInvitationCodeReq:
    properties:
      codes:
        items:
          type: string
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.DelDefaultFriendReq:
    properties:
      userIDs:
        items:
          type: string
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.DelDefaultGroupReq:
    properties:
      groupIDs:
        items:
          type: string
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.DelInvitationCodeReq:
    properties:
      codes:
        items:
          type: string
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.GenInvitationCodeReq:
    properties:
      chars:
        type: string
      len:
        type: integer
      num:
        type: integer
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.InvitationRegister:
    properties:
      createTime:
        type: integer
      invitationCode:
        type: string
      usedUser:
        $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_common.UserPublicInfo'
      usedUserID:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.SearchInvitationCodeReq:
    properties:
      codes:
        items:
          type: string
        type: array
      keyword:
        type: string
      pagination:
        $ref: '#/definitions/sdkws.RequestPagination'
      status:
        type: integer
      userIDs:
        items:
          type: string
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchInvitationCodeResp:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.InvitationRegister'
        type: array
      total:
        type: integer
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_chat.RegisterUserInfo:
    properties:
      RegisterType:
        type: integer
      account:
        type: string
      areaCode:
        type: string
      birth:
        type: integer
      email:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      nickname:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      userID:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_common.UserPublicInfo:
    properties:
      account:
        type: string
      email:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      level:
        type: integer
      nickname:
        type: string
      userID:
        type: string
    type: object
  sdkws.RequestPagination:
    properties:
      pageNumber:
        type: integer
      showNumber:
        type: integer
    type: object
...
host: localhost:10009
info:
  contact:
    email: ilovecoding@foxmail.com
    name: 飞宏IM技术支持
    url: https://github.com/liony823/chat
  description: 飞宏IM后台管理接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 飞宏IM后台管理接口
  version: 1.0.0
paths:
...
  /default/group/add:
    post:
      consumes:
      - application/json
      description: 添加用户注册时自动加入的默认群组
      operationId: addDefaultGroup
      parameters:
      - description: 默认群组信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddDefaultGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加默认群组
      tags:
      - default
  /default/group/del:
    post:
      consumes:
      - application/json
      description: 删除用户注册时自动加入的默认群组
      operationId: delDefaultGroup
      parameters:
      - description: 要删除的群组ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelDefaultGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除默认群组
      tags:
      - default
...
  /default/user/add:
    post:
      consumes:
      - application/json
      description: 添加用户注册时的默认好友
      operationId: addDefaultFriend
      parameters:
      - description: 默认好友信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddDefaultFriendReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加默认好友
      tags:
      - default
  /default/user/del:
    post:
      consumes:
      - application/json
      description: 删除用户注册时的默认好友
      operationId: delDefaultFriend
      parameters:
      - description: 要删除的好友ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelDefaultFriendReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除默认好友
      tags:
      - default
...
  /invitation_code/add:
    post:
      consumes:
      - application/json
      description: 手动添加新的邀请码
      operationId: addInvitationCode
      parameters:
      - description: 邀请码信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddInvitationCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加邀请码
      tags:
      - invitation
  /invitation_code/del:
    post:
      consumes:
      - application/json
      description: 删除指定的邀请码
      operationId: delInvitationCode
      parameters:
      - description: 要删除的邀请码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelInvitationCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除邀请码
      tags:
      - invitation
  /invitation_code/gen:
    post:
      consumes:
      - application/json
      description: 批量生成新的邀请码
      operationId: genInvitationCode
      parameters:
      - description: 生成邀请码的配置信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.GenInvitationCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 生成邀请码
      tags:
      - invitation
  /invitation_code/search:
    post:
      consumes:
      - application/json
      description: 根据条件搜索邀请码
      operationId: searchInvitationCode
      parameters:
      - description: 搜索条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchInvitationCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchInvitationCodeResp'
              type: object
      summary: 搜索邀请码
      tags:
      - invitation
...
  /user/import/json:
    post:
      consumes:
      - application/json
      description: 通过JSON数据批量导入用户信息
      operationId: importUserByJson
      parameters:
      - description: 用户信息列表
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.RegisterUserInfo'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 通过JSON导入用户
      tags:
      - import
  /user/import/xlsx:
    get:
      description: 下载用户导入的Excel模板文件
      operationId: importTemplate
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel模板文件
          schema:
            type: file
      summary: 获取导入模板
      tags:
      - import
    post:
      consumes:
      - multipart/form-data
      description: 通过Excel文件批量导入用户信息
      operationId: importUserByXlsx
      parameters:
      - description: Excel文件
        in: formData
        name: data
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 通过Excel导入用户
      tags:
      - import
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: '"Type ''Bearer'' followed by a space and JWT token"'
    in: header
    name: token
    type: apiKey
swagger: "2.0"